Return collection list without resource with lookup problems
When a collection receives a GET request, it is going to fetch all
resources names, in a list,  and perform a lookup for each resource,
building the return list with all data.
There is a potencial race condition problem in this approach. If a
given resource is removed between list of names creation and its
lookup call, wok is going to fail and the exception is not treated.
This patch fix this problem catching the exception, logging the problem
and completing the GET request. So, the resource with problem is
discarted and other resources are returned.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>