Sort user request log output
There is an issue in user request log output when you have backup files
in your system. The values from backup logs are included in the return
list and showed before latest events. Then user has to use UI sort
buttons in order to see what is happening.
This patch sorts the output, ensuring that latest events will be
returned on 'top' of the list and be showed firstly in the UI.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>