UI: Create notification-container and make notifications persistent
There is an issue in frontend that does not allow notification messages
to be showed more than once if the problem happens again.
Also, when notifications are created, they are showed in different parts
of the UI, for each plugin or plugin tab (usually in the alert-container).
If some window does not have alert-container, the notification is not
showed.
This patch creates a notificatoin-container 'on-the-fly' in the windows
that Wok is displaying and shows the notification.
Notification is check by its message content, instead of the code ID,
allowing multiple similar notifications but with different subjects
(like different VM names).
Notifications are showed again if message was closed but backend returns
the same notification massage. Before, messages were nevermore display.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>