Bug fix #173: Get color tab from tab-ext.xml file
The color tab must be specified by plugin instead of hardcoded on Wok
source code.

So 2 atrributes were added to <functionality> tag: colorTab1 for the
first tab level color and colorTab2 for the second tab level color.

This patch also removes CSS configuration related to the navigation bar as it
is now set on JS per plugin configuration.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>