Bug fix #173: Get plugin color to set user log data
The user log activity lists all the actions done on Wok and its plugins.
To better represent the action for a given plugin, the plugin color is
display in addition to the plugin name.

As plugin color is now specified on tab-ext.xml, store the value per
plugin to be used on user log UI.

This patch also removes CSS configuration used to set the plugin color.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>