Generate SSL self signed certificate on package post installation script
When Wok is installed from a package, the SSL self signed certificate should be
created on package post installation script instead of on server starts
up.

That will be required to fix issue #175. As the nginx configuration will
not be created on the fly, as the SSL cert/key must be there to do not make
nginx fails.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>