Add apache configuration
Provide an apache2 virtual host configuration in contrib. With this,
the user can choose to use either nginx or apache as a proxy.

The reason why the configuration has been added to contrib rather
than src/apache2 to mimic the nginx config, is that there is no distro
with the same apache2 config layout than others.

Also add Wants and After for apache2 in the systemd unit file.

Signed-off-by: Cédric Bosdonnat <cbosdonnat@suse.com>