Removing sessionStorage call in wok.main.js
sessionStorage was being used in early versions of the websocket
notifications implementation to avoid opening multiple connections,
before the current single permanent connection approach was
implemented. If the 'wokNotificationWebSocket' key existed in the
sessionStorage, we skipped the opening of a new websocket.

This patch removes the sessionStorage usage because it accomplishes
nothing in current code. No verification is made using the created
'wokNotificationWebSocket' session key in the code.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>